// payload-types.ts
/* tslint:disable */
/**
 * This file was automatically generated by Payload CMS.
 * DO NOT MODIFY IT BY HAND. Instead, modify your source Payload config,
 * and re-run `payload generate:types` to regenerate this file.
 */

export interface Config {
  collections: {
    pages: Page;
    media: Media;
    products: Product;
    colors: Color;
    sizes: Size;
    users: User;
  };
  globals: {};
}
export interface Page {
  id: string;
  title: string;
  image?: string | Media;
  layout?: (
    | {
        content?: {
          [k: string]: unknown;
        }[];
        buttons: {
          label: string;
          type: 'page' | 'custom';
          page: string | Page;
          url: string;
          newTab: boolean;
          id?: string;
        }[];
        id?: string;
        blockName?: string;
        blockType: 'cta';
      }
    | {
        content?: {
          [k: string]: unknown;
        }[];
        id?: string;
        blockName?: string;
        blockType: 'content';
      }
    | {
        image: string | Media;
        type?: 'normal' | 'fullscreen' | 'wide';
        caption?: {
          [k: string]: unknown;
        }[];
        id?: string;
        blockName?: string;
        blockType: 'image';
      }
  )[];
  meta?: {
    title?: string;
    description?: string;
    keywords?: string;
  };
  slug?: string;
  updatedAt: string;
  createdAt: string;
}
export interface Media {
  id: string;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string;
  filename?: string;
  mimeType?: string;
  filesize?: number;
  width?: number;
  height?: number;
  sizes?: {
    card?: {
      url?: string;
      width?: number;
      height?: number;
      mimeType?: string;
      filesize?: number;
      filename?: string;
    };
    feature?: {
      url?: string;
      width?: number;
      height?: number;
      mimeType?: string;
      filesize?: number;
      filename?: string;
    };
  };
}
export interface Product {
  id: string;
  name?: string;
  price?: number;
  colors?: string[] | Color[];
  sizes?: string[] | Size[];
}
export interface Color {
  id: string;
  name?: string;
}
export interface Size {
  id: string;
  name?: string;
}
export interface User {
  id: string;
  updatedAt: string;
  createdAt: string;
  email?: string;
  resetPasswordToken?: string;
  resetPasswordExpiration?: string;
  loginAttempts?: number;
  lockUntil?: string;
  password?: string;
}